Fix/clang compiler errors (#283)
* Fix clang compiler missing override warning in pilz_joint_trajectory_controller

* Fix comma problem with INSTANTIATE_TEST_CASE_P
There exists a compiler warning named "-Wgnu-zero-variadic-macro-arguments".
If build with gcc this warning does not appear, however, build with clang
the compiler warning appears. To avoid this problem a comma is added which
resolves the problem.

* Fix implicit conversion errors

+++ prbt_hardware_support/test/unittests/unittest_speed_observer.cpp
@@ -141,21 +141,21 @@ MATCHER_P(StoState, x, "Sto state " + std::string(negation ? "is not" : "is") + 
 MATCHER_P2(NameAtI, i, name,
            "Name at index " + PrintToString(i) + std::string(negation ? "is not" : "is") + ": " + name + ".")
 {
-  return static_cast<unsigned long>(i) < arg->name.size() && arg->name.at((unsigned)i).compare(name) == 0;
+  return static_cast<size_t>(i) < arg->name.size() && arg->name.at(static_cast<size_t>(i)).compare(name) == 0;
 }
 
 MATCHER_P2(SpeedAtIGe, i, x,
            "Speed at index " + PrintToString(i) + std::string(negation ? "is not" : "is") + " greater or equal to" +
            PrintToString(x) + ".")
 {
-  return static_cast<unsigned long>(i) < arg->name.size() && arg->speed.at(i) >= x;
+  return static_cast<size_t>(i) < arg->name.size() && arg->speed.at(static_cast<size_t>(i)) >= x;
 }
 
 MATCHER_P2(SpeedAtILe, i, x,
            "Speed at index " + PrintToString(i) + std::string(negation ? "is not" : "is") + " less or equal to" +
            PrintToString(x) + ".")
 {
-  return static_cast<unsigned long>(i) < arg->name.size() && arg->speed.at(i) <= x;
+  return static_cast<size_t>(i) < arg->name.size() && arg->speed.at(static_cast<size_t>(i)) <= x;
 }
 
 /**
